setup unit testing of sql queries
